Trait serde_generate::SourceInstaller

pub trait SourceInstaller {
    type Error;
    fn install_module(
        &self, 
        config: &CodeGeneratorConfig, 
        registry: &Registry
    ) -> Result<(), Self::Error>;

    fn install_serde_runtime(&self) -> Result<(), Self::Error>;

    fn install_bincode_runtime(&self) -> Result<(), Self::Error>;

    fn install_lcs_runtime(&self) -> Result<(), Self::Error>;
}

How to copy generated source code and available runtimes for a given language.

Associated Types

type Error

Loading content...
Required methods

fn install_module(
&self,
config: &CodeGeneratorConfig,
registry: &Registry
) -> Result<(), Self::Error>

Create a module exposing the container types contained in the registry.

fn install_serde_runtime(&self) -> Result<(), Self::Error>

Install the serde runtime.

fn install_bincode_runtime(&self) -> Result<(), Self::Error>

Install the bincode runtime.

fn install_lcs_runtime(&self) -> Result<(), Self::Error>

Install the Libra Canonical Serialization (LCS) runtime.

Loading content...